+++ test/modules/topics/components/menu/catalog/CatalogLeaf.test.js
@@ -8,6 +8,7 @@ import { createDefaultLayerProperties } from '../../../../../../src/store/layers
 import { WMTSGeoResource } from '../../../../../../src/services/domain/geoResources';
 import { Checkbox } from '../../../../../../src/modules/commons/components/checkbox/Checkbox';
 import { modalReducer } from '../../../../../../src/store/modal/modal.reducer';
+import { TemplateResult } from 'lit-html';
 
 
 
@@ -174,6 +175,7 @@ describe('CatalogLeaf', () => {
 					icon.click();
 
 					expect(store.getState().modal.data.title).toBe('someLabel');
+					expect(store.getState().modal.data.content).toBeInstanceOf(TemplateResult);
 				});
 			});
 		});
